+++ packages/ice/src/types/userConfig.ts
@@ -13,9 +13,15 @@ interface SyntaxFeatures {
 
 interface Optimization {
   /**
-   * Optimize code by remove react-router dependencies when set to true.
+   * Optimize code by remove react-router dependencies when set to true,
+   * it only works when route count is 1.
    */
   router?: boolean;
+  /**
+   * @private
+   * Remove react-router dependencies by force, even if route count is greater than 1.
+   */
+  disableRouter?: boolean;
 }
 
 interface MinifyOptions {
@@ -29,6 +35,8 @@ interface IgnorePattern {
 }
 
 type DistType = 'javascript' | 'html';
+
+type DropType = 'trace' | 'debug' | 'log' | 'info' | 'warn' | 'error';
 
 interface Fetcher {
   packageName: string;
@@ -140,7 +148,7 @@ export interface UserConfig {
    * `console.*` will be dropped when build.
    * @see https://v3.ice.work/docs/guide/basic/config#droploglevel
    */
-  dropLogLevel?: 'trace' | 'debug' | 'log' | 'info' | 'warn' | 'error';
+  dropLogLevel?: DropType[] | DropType | boolean;
   /**
    * Minify build output, it only works in prod mode by default.
    * @see https://v3.ice.work/docs/guide/basic/config#minify
@@ -243,14 +251,16 @@ export interface UserConfig {
    * Code splitting strategy, support page and vendors, default value is true (built-in strategy).
    * @see https://v3.ice.work/docs/guide/basic/config#codesplitting
    */
-  codeSplitting?: 'page' | 'vendors' | boolean;
+  codeSplitting?: 'page' | 'vendors' | 'page-vendors' | boolean;
   /**
    * generate additional assets for request data, default is true
    * @see https://v3.ice.work/docs/guide/basic/config#dataloader
    */
-  dataLoader?: {
-    fetcher?: Fetcher;
-  } | Boolean;
+  dataLoader?:
+    | {
+        fetcher?: Fetcher;
+      }
+    | Boolean;
   /**
    * Enable cross-origin loading of chunks.
    * @see https://v3.ice.work/docs/guide/basic/config#crossoriginloading