+++ packages/ice/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 3.2.8
+
+### Patch Changes
+
+- 27a72536: fix: always external dependencies when get config
+- cf8a78e3: feat: support code splitting strategy of page-vendors
+- b691b9e9: feat: support remove router even if route count is greater than 1
+- Updated dependencies [cf8a78e3]
+- Updated dependencies [b691b9e9]
+- Updated dependencies [92ec144f]
+  - @ice/webpack-config@1.0.19
+  - @ice/runtime@1.2.6
+
 ## 3.2.7
 
 ### Patch Changes